Change index in Peek and Poke node to uint16

+++ tools.go
@@ -4,7 +4,7 @@ import (
 	"github.com/anon55555/mt"
 )
 
-func (mb *MapBlk) PeekNode(i int) mt.Node {
+func (mb *MapBlk) PeekNode(i uint16) mt.Node {
 	return mt.Node{
 		Param0: mb.Param0[i],
 		Param1: mb.Param1[i],
@@ -12,7 +12,7 @@ func (mb *MapBlk) PeekNode(i int) mt.Node {
 	}
 }
 
-func (mb *MapBlk) PokeNode(i int, node mt.Node) {
+func (mb *MapBlk) PokeNode(i uint16, node mt.Node) {
 	mb.Param0[i] = node.Param0
 	mb.Param1[i] = node.Param1
 	mb.Param2[i] = node.Param2
